refact. Form.tsx (to common)
